+++ src/Enums/Format.php
@@ -1,5 +1,14 @@
 <?php
 
+/*
+ * This file is part of the 2amigos/qrcode-library project.
+ *
+ * (c) 2amigOS! <http://2am.tech/>
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE file that was distributed with this source code.
+ */
+
 namespace Da\QrCode\Enums;
 
 use Da\QrCode\Format\BookMarkFormat;
@@ -15,9 +24,8 @@ use Da\QrCode\Format\SmsFormat;
 use Da\QrCode\Format\VCardFormat;
 use Da\QrCode\Format\WifiFormat;
 use Da\QrCode\Format\YoutubeFormat;
-use MabeEnum\Enum;
 
-final class Format extends Enum
+final class Format
 {
     public const TEXT = 'text';
     public const BOOK_MARK = BookMarkFormat::class;