+++ src/Label.php
@@ -46,7 +46,7 @@ class Label implements LabelInterface
      * @param string|null $alignment
      * @param array       $margins
      */
-    public function __construct(string $text, string $font = null, $fontSize = null, $align = null, array $margins = [])
+    public function __construct(string $text, ?string $font = null, ?int $fontSize = null, ?string $align = null, array $margins = [])
     {
         $this->text = $text;
         $this->font = $font ?: __DIR__ . '/../resources/fonts/noto_sans.otf';